third_party/WebKit/LayoutTests/fast/css/font-shorthand-from-longhands-expected.txt

Test the return values for the font properties on the style object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.font is ''
PASS computedStyle.font is 'normal normal 400 normal 20px / normal foobar'
PASS computedStyle.fontSize is '20px'
PASS checkFontStyleValue() is true
PASS style.font is ''
PASS computedStyle.font is 'normal normal 400 normal 20px / normal sans-serif'
PASS computedStyle.fontFamily is 'sans-serif'
PASS style.font is ''
PASS computedStyle.font is 'italic normal 400 normal 20px / normal sans-serif'
PASS computedStyle.fontStyle is 'italic'
PASS style.font is ''
PASS computedStyle.font is 'italic small-caps 400 normal 20px / normal sans-serif'
PASS computedStyle.fontVariant is 'small-caps'
PASS style.font is ''
PASS computedStyle.font is 'italic small-caps 700 normal 20px / normal sans-serif'
FAIL computedStyle.fontWeight should be bold. Was 700.
PASS style.font is ''
PASS computedStyle.font is 'italic small-caps 700 normal 20px / 40px sans-serif'
PASS computedStyle.lineHeight is '40px'
PASS style.font is 'italic small-caps bold ultra-expanded 20px/40px sans-serif'
PASS computedStyle.font is 'italic small-caps 700 ultra-expanded 20px / 40px sans-serif'
PASS checkFontStyleValue() is true
PASS style.font is ''
PASS computedStyle.font is ''
PASS style.font is 'italic small-caps bold ultra-expanded 20px/40px sans-serif'
PASS computedStyle.font is 'italic small-caps 700 ultra-expanded 20px / 40px sans-serif'
PASS style.font is ''
PASS computedStyle.font is ''
PASS style.font is 'italic small-caps bold ultra-expanded 20px/40px sans-serif'
PASS computedStyle.font is 'italic small-caps 700 ultra-expanded 20px / 40px sans-serif'
PASS style.font is ''
PASS computedStyle.font is 'normal normal 400 normal 16px / normal foobar'
PASS checkFontStyleValue() is true
PASS style.fontVariantCaps is 'all-small-caps'
PASS computedStyle.font is ''
PASS style.font is ''
PASS successfullyParsed is true

TEST COMPLETE
